+++ include/reMatrix3.h
@@ -1,7 +1,7 @@
 //////////////////////////////////////////////////////////////////////////////////////////////////
 //
 // File:        reMatrix3.h
-// Project:     Razor Edge Classes
+// Project:     reMath
 // Description: Definition of Matrix3 class (Math Module)
 // Copyright:   Copyright © 2004++ REGLabs
 // Author:      Pavel Chikul