+++ include/reVec3d.h
@@ -1,7 +1,7 @@
 //////////////////////////////////////////////////////////////////////////////////////////////////
 //
 // File:        reVec3d.h
-// Project:     Razor Edge Classes
+// Project:     reMath
 // Description: Definition of Vec3d class (Math Module)
 // Copyright:   Copyright © 2004++ REGLabs
 // Author:      Pavel Chikul
@@ -40,6 +40,9 @@ namespace re
 
 		// Get vector magnitutde.
 		float length() const;
+
+		// Calculate absolute distance to another vector.
+		float distanceTo(const Vec3d& vector) const;
 
 		// Parallel vectors check.
 		bool isParallel(const Vec3d& vector) const;